navbar component done
