+++ md.safin-portfolio/src/App.js
@@ -5,101 +5,7 @@ function App() {
   return (
     <div className="App">
       <Navbar />
-      <header className="App-header">
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-        efsesfsefsefsefsfsefsffnlksdndnsdSGsdgsdgsg
-      </header>
+      <header className="App-header"></header>
     </div>
   );
 }
